Updated header image and altered styling of site title, sub-title.
Was unhappy with pixilation in previous site header image. Grabbed new slice of source
image that was wide enough to not be expanded when shown on the site.

Changed text color for title and sub-title to black so text would be visible now that it
is against lighter (white) portion of header image.